+++ core/classes/actions/class.action.switchPhpExtension.php
@@ -14,14 +14,14 @@ class ActionSwitchPhpExtension
             $offContent = ';extension=' . $args[0];
             $extExists = file_exists($neardBins->getPhp()->getExtPath() . '/' . $args[0] . '.dll');
             
-            $phpiniContent = file_get_contents($neardBins->getPhp()->getApacheConf());
+            $phpiniContent = file_get_contents($neardBins->getPhp()->getConf());
             if ($args[1] == self::SWITCH_ON) {
                 $phpiniContent = str_replace($offContent, $onContent, $phpiniContent);
             } elseif ($args[1] == self::SWITCH_OFF) {
                 $phpiniContent = str_replace($onContent, $offContent, $phpiniContent);
             }
             
-            $phpiniContentOr = file_get_contents($neardBins->getPhp()->getApacheConf());
+            $phpiniContentOr = file_get_contents($neardBins->getPhp()->getConf());
             if ($phpiniContent == $phpiniContentOr && $extExists) {
                 $extsIni = $neardBins->getPhp()->getExtensionsFromIni();
                 $latestExt = (end($extsIni) == '0' ? ';' : '') . 'extension=' . key($extsIni);
@@ -32,7 +32,7 @@ class ActionSwitchPhpExtension
                 );
             }
             
-            file_put_contents($neardBins->getPhp()->getApacheConf(), $phpiniContent);
+            file_put_contents($neardBins->getPhp()->getConf(), $phpiniContent);
         }
     }